backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.37
psycopg[binary]==3.2.3
redis==5.2.1
pydantic-settings==2.8.0
python-dateutil==2.9.0.post0

